pyproject.toml
# ==================================================================================================
# Pyproject Configuration for the Glossa Package
# ==================================================================================================

# --- Build System Requirements --------------------------------------------------------------------

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


# --- Package Metadata -----------------------------------------------------------------------------

[project]
name = "glossa"
version = "0.0.0"
authors = [{name = "Esther Poniatowski", email = "esther.poniatowski@ens.psl.eu"}]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
readme = {file = "README.md", content-type = "text/markdown"}
description = "Documentation linter for Python projects"
keywords = ["documentation", "linter", "docstrings", "quality", "static-analysis"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha"
]
urls = {homepage = "https://github.com/esther-poniatowski/glossa"}


# --- Dependencies ---------------------------------------------------------------------------------

requires-python = ">=3.12"

dependencies = [
    "typer",                 # CLI interface
    "rich",                  # Terminal output formatting
    "pyyaml>=6.0",           # YAML configuration support
]


# --- Package Source Code --------------------------------------------------------------------------

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# --- CLI Entry Points -----------------------------------------------------------------------------

[project.scripts]
glossa = "glossa.adapters.cli:app"

# --- Tool Configurations --------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
